Move test namedcache factory to test utils
Fix cordapp loader tests
