+++ node/src/main/kotlin/net/corda/node/internal/cordapp/JarScanningCordappLoader.kt
@@ -10,6 +10,7 @@ import net.corda.core.internal.*
 import net.corda.core.internal.cordapp.CordappImpl
 import net.corda.core.internal.cordapp.CordappInfoResolver
 import net.corda.core.internal.notary.NotaryService
+import net.corda.core.internal.notary.TrustedAuthorityNotaryService
 import net.corda.core.node.services.CordaService
 import net.corda.core.schemas.MappedSchema
 import net.corda.core.serialization.SerializationCustomSerializer
@@ -128,7 +129,8 @@ class JarScanningCordappLoader private constructor(private val cordappJarPaths: 
     }
 
     private fun findNotaryService(scanResult: RestrictedScanResult): Class<out NotaryService>? {
-        val result = scanResult.getClassesWithSuperclass(NotaryService::class)
+        val result = scanResult.getClassesWithSuperclass(NotaryService::class) +
+                scanResult.getClassesWithSuperclass(TrustedAuthorityNotaryService::class)
         logger.info("Found notary service CorDapp implementations: " + result.joinToString(", "))
         return result.firstOrNull()
     }
